requirements.txt
python-telegram-bot==21.6
anthropic==0.40.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
APScheduler==3.10.4
python-dateutil==2.9.0
